Delete test scenario bug removed

+++ app/Http/Controllers/TestScenarioController.php
@@ -203,9 +203,16 @@ class TestScenarioController extends Controller {
 				Toast::message($message, 'danger');
 			}  else{      	
 	        	//Delete testcase
-	        	$item->delete();
-				$message = $this->getMessage('messages.delete_success');
-				Toast::success($message);
+	        	$del_res = $del_obj->deleteScLabByScenarioId($id);
+				if($del_res == 0)
+				{
+					$message = $this->getMessage('messages.delete_failed');
+					Toast::message($message, 'danger');
+				}  else{  
+		        	$item->delete();
+					$message = $this->getMessage('messages.delete_success');
+					Toast::success($message);
+				}
 			}
         }
 		else{
